refactor: Event delete logic is now inside deleteModal file. Also, events now persist to user even if they don't have an events array yet.
